backend/requirements.txt
fastapi
uvicorn
soundcard
numpy
deepgram-sdk
python-dotenv
websockets
openai
groq
firebase-admin
